drm/vc4: Set up an HVS display list so that we can scan out our FB.
The FB mailbox command we were using can't take in a physical address
-- it wants to allocate the memory itself out of the area that the VPU
controls.  So, we use the FB command to get the mode set done, then we
smash the HVS display list that was set up to actually scan out of our
buffer.

Now the KMS console actually comes up.

Signed-off-by: Eric Anholt <eric@anholt.net>