drm/vc4: Use interrupts for binner overflow memory allocation handling.
This avoids some of the checks in the spinloop waiting for frame done,
and avoids the race I had documented, as far as I know.

Note that this reportedly requires "mask_gpu_interrupt0=0x400" in
config.txt to keep the VPU firmware from trying to do things based on
our interrupts.

Signed-off-by: Eric Anholt <eric@anholt.net>