drm/vc4: Move the hang detection to a timer.
We're going to need to do this for async execution, so make the switch
now.

Since the detection is smarter now (it can see when you're making
progess through the lists), we can crank up the default timeout for
executing a job.

v2: Move vc4_gem_init() out from in between drm_irq_install() and its
    error check.

Signed-off-by: Eric Anholt <eric@anholt.net>