vc4: Add a BO cache for freed BOs.
The kernel needs to allocate BOs to execute rendering (both the exec
BO to store the user's commands, and the binner overflow BO to handle
the hardware's own needs for memory).  We need to cache these, because
getting a new CMA allocation is very expensive.

Note that userspace has its own BO cache because it wants to avoid the
cost of mmaping BOs in addition to the cost of allocation.

This is about a 15% performance improvement on glxgears under X, which
is thoroughly CPU limited and not hitting refresh rate yet.  Next step
is to cache shader validation (half of the ioctl overhead) and avoid
unnecessary depth writes (a big chunk of memory bandwidth overhead).

Signed-off-by: Eric Anholt <eric@anholt.net>