drm/vc4: Drop the mbox mode set command.
It seems that the firmware ends up triggering a reset of our display
lists later if we send this command.  Significantly later.  Instead,
just rely on the firmware already having set up a video mode before
handing off to the linux kernel, and just rewrite the existing display
lists.

This is a terrible hack, but at least it's a terrible hack that seems
to work now.

Signed-off-by: Eric Anholt <eric@anholt.net>