drm/vc4: Add a delayed fixup of the display lists.
Sometimes the firmware will come along and rewrite them on us.  Until
we get a firmwrare that we can tell to quick screwing with our
display, just try to fix it up periodically.

The need for fixups can at least be reduced with "avoid_warnings=2" in
the config.txt, but I don't think that will be fully reliable.

Signed-off-by: Eric Anholt <eric@anholt.net>