clk: Add a Raspberry Pi-specific clock driver.
Unfortunately, the clock manager's registers are not accessible by the
ARM, so we have to request that the firmware modify our clocks for us.

This driver only registers the clocks at the point they are requested
by a client driver.  This is partially to support returning
-EPROBE_DEFER when the firmware driver isn't supported yet, but it
also avoids issues with disabling "unused" clocks due to them not yet
being connected to their consumers in the DT.

v2: Declare the mutex static (from review by Baruch Siach), merge
    description and copyright comments.

v3: Update for new rpi firmware API.  Update the compatible string.
    Make the firmware handle be under a vendor-namespaced property.
    Make the DT indices match the firmware clock IDs.  Move the driver
    under the firmware driver's Kconfig, since it requires it.  Move a
    container_of() from 2 callers into the callee.

v4: Update commit message to mention subsystem.  Add #defines for a
    couple of bitfield checks, and use the bitfield check for
    rpi_clk_is_on().  Return an error when appropriate in
    rpi_clk_is_on(), and don't return specific rpi_firmware_property()
    errors in other places when we were only returning -EINVAL.  Add
    some comment text describing how return values work from
    the firmware.

v5: Restructure as a platform_driver on suggestion from Michael
    Turquette.  This works around the -EPROBE_DEFER problem, so now we
    can register all of our clocks at probe time.  Store the current
    clock rate in the struct, and avoid asking for no-op changes.  It
    appears that a no-op change of the pixel clock while the clock is
    being used (at boot time when the KMS driver calls
    clk_prepare_enable()) actually breaks the display.  Dropped
    Stephen's ack since this is a pretty significant rewrite.

Signed-off-by: Eric Anholt <eric@anholt.net>