drm/vc4: Don't bother flushing the ARM CPU caches on GEM exec.
All of our BOs are mapped write-combining both in the kernel and
userspace by drm_gem_cma_helper.c, so anything we've written will end
up in main memory, anyway.  The only coherency trouble we might have
would be if we had content in the CPU's write combining buffers that
hadn't been flushed, but a CPU cache flush is unlikely to help with
that, anyway.

Improves no-swapbuffers drawarrays-mode isosurf performance by around 20%.

Signed-off-by: Eric Anholt <eric@anholt.net>