drm/vc4: Prevent clients from racing far ahead of execution.
This is policy that we've often left up to userland in various DRM
drivers.  However, given how thoroughly memory-constrained we are on
VC4, letting userspace run away with things and over-inflate their
caches is definitely not helping.

This hasn't reduced maximum throughput in my testing, since we're only
waiting until the exec just before the one we queued is finished, so
there will be a job to submit the moment the GPU sends us the frame
done interrupt.  We would only actually idle the GPU if assembling
some scenes takes more CPU time than the time it takes to complete
some scenes on the GPU.

Signed-off-by: Eric Anholt <eric@anholt.net>